fix: incorrect TadoX base class
